requirements.txt
streamlit
pandas
openai
geopy
folium
streamlit-folium
pydantic
python-dotenv
watchdog